requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
fastparquet==2023.10.1
xlrd==2.0.1